fix(support): fix http request to api service
